Implement ColumnReader
This commit completes the implementation for ColumnReader, which can be
used to read definition levels, repetition levels, and values from a row
group. Currently it only support 3 encoding types: PLAIN and PLAIN/RLE
DICTIONARY (the latter two can actually be considered as one). Also
added unit tests for the column reader, but it currently only covers i32
and i64 types. These mostly follow the C++ implementation.

This commit also bundles a few other changes that are relatively minor:
1. Changed the type for 'bit_width' from usize to u8. The range for
   bit_width is always [0, 64] so no need to use usize.
2. Added 'get_column_reader()' and 'get_typed_column_reader()'. The
   first fn is used to produce a column reader given a specific column
   descriptor, while the second is cast a specific column reader to a
   generic column reader on a type 'T'.

+++ src/encodings/decoding.rs
@@ -273,7 +273,7 @@ impl<T: DataType> DictDecoder<T> {
 impl<T: DataType> Decoder<T> for DictDecoder<T> {
   fn set_data(&mut self, data: ByteBufferPtr, num_values: usize) -> Result<()> {
     // first byte in `data` is bit width
-    let bit_width = data.as_ref()[0] as usize;
+    let bit_width = data.as_ref()[0];
     let mut rle_decoder = RleDecoder::new(bit_width);
     rle_decoder.set_data(data.start_from(1));
     self.num_values = num_values;
